Add preliminary cache feature
This uses an ES2015 Map to cache every result behind the key it was loaded for.

There's no way to configure or disable the caching logic, but that
could be a cool thing to do in the future.

This also adds test cases to assert that values are cached and not passed
to the batch function when cached.