Add loadMany function, refactor so that fns can reference eachother
This adds a loadMany function to match the reference implementation.

This includes a refactor, since loadMany just calls load with all of
the keys given and then wraps those in a single Promise.all call.

In order to not use `this`, this moves the function definitions
outside of the object literal return value, and names them so they can
be referenced by other functions in the public API.